C++ program that processes infinitely big numbers
Uses a List ADT to store and retrieve the numbers. The List ADT has 3
different implementation that the user can choose by changing a line on
the constructor of LongInteger.cpp. A step of the Karatsuba algorithm
was use on the multiplication implementation while Exponentiation by
Squaring was used for the Power implementation.